+++ bpod_core/bpod/constants.py
@@ -50,7 +50,7 @@ CHANNEL_TYPES_OUTPUT.update(
     }
 )
 N_SERIAL_EVENTS_DEFAULT = 15
-VALID_OPERATORS = {'exit', '>exit', '>back'}
+VALID_OPERATORS = {'>exit', '>back'}
 MACHINE_TYPES = {3: 'r2.0-2.5', 4: '2+ r1.0'}
 CONFIG_PATH = platformdirs.user_config_path(appname='bpod-core', appauthor=False)
 DISCOVERY_TIMEOUT = 0.11